requirements.txt
Django==4.1
pytest==7.1.2
factory-boy==3.2.1
django-filter==22.1
djangorestframework==3.13.1
django-extensions==3.2.0
Markdown==3.4.1
requests==2.28.1
django-debug-toolbar==3.6.0
psycopg2-binary==2.9.3
flake8==5.0.4